run e2e in angular 6 too; check log file

+++ e2e.sh
@@ -14,8 +14,11 @@ npm pack
 
 npm i @angular/cli -g
 
+# goto e2e directory
+cd ../../e2e$
+$ echo $PWD
+
 # install packages & generate electron project
-cd ../../e2e
 npm i ../packages/schematics/electron-schematics-schematics-0.0.1.tgz
 ng g @electron-schematics/schematics:electron
 npm i ../packages/build-electron/electron-schematics-build-electron-0.0.1.tgz
@@ -25,9 +28,9 @@ cp 'main.ts' 'projects/electron/main.ts'
 ng serve electron
 
 # check if log file exists
-# if [[ ! -f e2e.log ]] ; then
-#     exit 1
-# fi
+if [[ ! -f e2e.log ]] ; then
+    exit 1
+fi
 
 # try to build the app
 rm -r ./dist
